requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
google-api-python-client==2.131.0
google-auth==2.29.0
